+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/section/call/precompileSubsection/PrecompileSubsection.java
@@ -156,7 +156,7 @@ public abstract class PrecompileSubsection
     }
 
     // successful PRC_CALL to MODEXP
-    final ModexpSubsection modexpSubsection = (ModexpSubsection) this;
+    final LondonModexpSubsection modexpSubsection = (LondonModexpSubsection) this;
     final int mbs = modexpSubsection.getForkAppropriateModexpMetadata().mbsInt();
     final int maxInputSize = modexpSubsection.getForkAppropriateModexpMetadata().getMaxInputSize();
     final Bytes returnData = frame.getReturnData();
